Move `TasksSubservice` to `TasksService` with tests and create `TasksModule`
