+++ nest-city-account/src/app.module.ts
@@ -8,10 +8,10 @@ import { PrismaModule } from './prisma/prisma.module'
 import { VerificationModule } from './user-verification/verification.module'
 import { UserModule } from './user/user.module'
 import AppLoggerMiddleware from './utils/middlewares/logger.service'
-import { TasksSubservice } from './utils/subservices/tasks.subservice'
 import ThrowerErrorGuard from './utils/guards/errors.guard'
 import { TaxSubservice } from './utils/subservices/tax.subservice'
 import { ConfigModule } from '@nestjs/config'
+import { TasksModule } from './tasks/tasks.module'
 
 @Module({
   imports: [
@@ -22,9 +22,10 @@ import { ConfigModule } from '@nestjs/config'
     AdminModule,
     ScheduleModule.forRoot(),
     ConfigModule.forRoot({ isGlobal: true }),
+    TasksModule,
   ],
   controllers: [AppController],
-  providers: [TaxSubservice, TasksSubservice, ThrowerErrorGuard],
+  providers: [TaxSubservice, ThrowerErrorGuard],
 })
 export class AppModule implements NestModule {
   configure(consumer: MiddlewareConsumer): void {
