+++ nest-city-account/src/tasks/__tests__/tasks.subservice.spec.ts
@@ -1,16 +1,17 @@
 import { createMock } from '@golevelup/ts-jest'
 import { Test, TestingModule } from '@nestjs/testing'
 import { Prisma, User } from '@prisma/client'
-import prismaMock from '../../../../test/singleton'
-import { AdminApi } from '../../../generated-clients/nest-tax-backend'
-import { PrismaService } from '../../../prisma/prisma.service'
-import { GdprSubType } from '../../../user/dtos/gdpr.user.dto'
-import ThrowerErrorGuard from '../../guards/errors.guard'
-import { DeliveryMethod } from '../../types/tax.types'
-import { TasksSubservice } from '../tasks.subservice'
-import { TaxSubservice } from '../tax.subservice'
+import prismaMock from '../../../test/singleton'
+import { AdminApi } from '../../generated-clients/nest-tax-backend'
+import { PrismaService } from '../../prisma/prisma.service'
+import { GdprSubType } from '../../user/dtos/gdpr.user.dto'
+import ThrowerErrorGuard from '../../utils/guards/errors.guard'
+import { DeliveryMethod } from '../../utils/types/tax.types'
+import { TasksService } from '../tasks.service'
+import { TaxSubservice } from '../../utils/subservices/tax.subservice'
+import { PhysicalEntityService } from '../../physical-entity/physical-entity.service'
 
-jest.mock('../../decorators/errorHandler.decorators', () => {
+jest.mock('../../utils/decorators/errorHandler.decorators', () => {
   return jest.fn(() => (target: object, propertyKey: string, descriptor: PropertyDescriptor) => {
     const originalMethod = descriptor.value
 
@@ -33,21 +34,22 @@ type UserWithRelations = Prisma.UserGetPayload<{
 }>
 
 describe('TasksSubservice', () => {
-  let service: TasksSubservice
+  let service: TasksService
 
   let throwerErrorGuard: ThrowerErrorGuard
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       providers: [
-        TasksSubservice,
+        TasksService,
         { provide: PrismaService, useValue: prismaMock },
         { provide: TaxSubservice, useValue: createMock<TaxSubservice>() },
+        { provide: PhysicalEntityService, useValue: createMock<PhysicalEntityService>()},
         ThrowerErrorGuard,
       ],
     }).compile()
 
-    service = module.get<TasksSubservice>(TasksSubservice)
+    service = module.get<TasksService>(TasksService)
     throwerErrorGuard = module.get<ThrowerErrorGuard>(ThrowerErrorGuard)
     Object.defineProperty(service, 'taxBackendAdminApi', { value: createMock<AdminApi>() }) // It is not injected so has to be set this way
   })
